Feature(RefactorNavBar)
