+++ client/app.js
@@ -1,14 +1,12 @@
 import React from 'react'
 
-import {Navbar} from './components'
-import Sidebar from './components/Sidebar/Sidebar.js'
+import { Navbar } from './components'
 import Routes from './routes'
 
 const App = () => {
   return (
     <div>
-      {/* <Navbar /> */}
-      <Sidebar />
+      <Navbar />
       <Routes />
     </div>
   )
