--- client/components/Sidebar/VisitorNav.js
@@ -1,43 +1,0 @@
-import React from 'react'
-import PropTypes from 'prop-types'
-import {connect} from 'react-redux'
-import {Link} from 'react-router-dom'
-import {logout} from '../../store'
-
-const Navbar = ({handleClick, isLoggedIn}) => (
-  <nav>
-    <div>
-    {/* The navbar will show these links before you log in */}
-    <Link to="/login">Login</Link>
-    <Link to="/signup">Sign Up</Link>
-  </div>
-</nav>
-
-)
-
-/**
- * CONTAINER
- */
-const mapState = state => {
-  return {
-    isLoggedIn: !!state.user.user.id
-  }
-}
-
-const mapDispatch = dispatch => {
-  return {
-    handleClick() {
-      dispatch(logout())
-    }
-  }
-}
-
-export default connect(mapState, mapDispatch)(Navbar)
-
-/**
- * PROP TYPES
- */
-Navbar.propTypes = {
-  handleClick: PropTypes.func.isRequired,
-  isLoggedIn: PropTypes.bool.isRequired
-}
