+++ public/style.css
@@ -12,7 +12,7 @@ label {
 
 nav {
   display: flex;
-  flex-direction: row;
+  /* flex-direction: row; */
   justify-content: space-between;
   align-items: flex-end;
 }
